+++ src/JavaCore/OOP/Phan1/VD30/mainMyDate.java
@@ -1,0 +1,12 @@
+package JavaCore.OOP.Phan1.VD30;
+
+public class mainMyDate {
+    public static void main(String[] args) {
+        myDate md = new myDate(15, 2, 2025);
+        md.printDate();
+        md.printDay();
+        md.printMonth();
+        md.printYear();
+        md.printYear();
+    }
+}
