+++ src/JavaCore/OOP/Phan1/VD31/Student/main_Student.java
@@ -1,0 +1,9 @@
+package JavaCore.OOP.Phan1.Student;
+
+public class main_Student {
+    public static void main(String[] args) {
+        Student studentClass = new Student();
+        studentClass.nhapThongTin();
+        studentClass.xuatThongTin();
+    }
+}
