--- src/JavaCore/OOP/VD30/mainMyDate.java
@@ -1,12 +1,0 @@
-package JavaCore.OOP.VD30;
-
-public class mainMyDate {
-    public static void main(String[] args) {
-        myDate md = new myDate(15, 2, 2025);
-        md.printDate();
-        md.printDay();
-        md.printMonth();
-        md.printYear();
-        md.printYear();
-    }
-}
