--- src/JavaCore/OOP/VD31/Student/main_Student.java
@@ -1,9 +1,0 @@
-package JavaCore.OOP.VD31.Student;
-
-public class main_Student {
-    public static void main(String[] args) {
-        Student studentClass = new Student();
-        studentClass.nhapThongTin();
-        studentClass.xuatThongTin();
-    }
-}
