--- src/JavaCore/OOP/VD36/HangSanXuat.java
@@ -1,27 +1,0 @@
-package JavaCore.OOP.VD36;
-
-public class HangSanXuat {
-    private String tenHangSanXuat;
-    private String quocGia;
-
-    public HangSanXuat(String tenHangSanXuat, String quocGia) {
-        this.tenHangSanXuat = tenHangSanXuat;
-        this.quocGia = quocGia;
-    }
-
-    public String getTenHangSanXuat() {
-        return tenHangSanXuat;
-    }
-
-    public void setTenHangSanXuat(String tenHangSanXuat) {
-        this.tenHangSanXuat = tenHangSanXuat;
-    }
-
-    public String getQuocGia() {
-        return quocGia;
-    }
-
-    public void setQuocGia(String quocGia) {
-        this.quocGia = quocGia;
-    }
-}
